backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic[email]>=2.5.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
